+++ src/main/java/com/example/finanzasgrupo5backend/Credito1/Model/MoraCredito1Response.java
@@ -14,7 +14,7 @@ public class MoraCredito1Response {
     private Long id;
     private String TEPm;
     private Double tasa;
-    private Long dias_atraso;
+    private Double dias_atraso;
     private Double total_moras;
 
     //private Credito1 credito1;
